start moving out wallet stuff
