+++ app/src/main/java/com/example/hashcache/AppHome.java
@@ -19,10 +19,6 @@ import android.widget.ImageButton;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.PopupMenu;
 
-import com.example.hashcache.database_connections.callbacks.GetPlayerCallback;
-import com.example.hashcache.models.Player;
-import com.example.hashcache.models.PlayerList;
-
 public class AppHome extends AppCompatActivity {
 
     @Override
