+++ app/src/main/java/com/example/hashcache/controllers/AddUserCommand.java
@@ -1,11 +1,8 @@
 package com.example.hashcache.controllers;
 
-import com.example.hashcache.database_connections.callbacks.BooleanCallback;
+import com.example.hashcache.models.database_connections.callbacks.BooleanCallback;
 import com.example.hashcache.models.PlayerList;
-import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
-
-import java.util.HashMap;
 
 public class AddUserCommand {
     FirebaseFirestore db;
