+++ app/src/main/java/com/example/hashcache/models/database_connections/FireStoreHelper.java
@@ -1,10 +1,10 @@
-package com.example.hashcache.database_connections.converters;
+package com.example.hashcache.models.database_connections;
 
 import android.util.Log;
 
 import androidx.annotation.NonNull;
 
-import com.example.hashcache.database_connections.callbacks.BooleanCallback;
+import com.example.hashcache.models.database_connections.callbacks.BooleanCallback;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
