+++ app/src/main/java/com/example/hashcache/models/database_connections/callbacks/GetPlayerCallback.java
@@ -1,4 +1,4 @@
-package com.example.hashcache.database_connections.callbacks;
+package com.example.hashcache.models.database_connections.callbacks;
 
 //Code from https://stackoverflow.com/a/48500679 add proper reference later
