+++ app/src/main/java/com/example/hashcache/models/database_connections/callbacks/GetScannableCodeCallback.java
@@ -1,4 +1,4 @@
-package com.example.hashcache.database_connections.callbacks;
+package com.example.hashcache.models.database_connections.callbacks;
 
 import com.example.hashcache.models.ScannableCode;
