+++ app/src/test/java/com/example/hashcache/stubs/PlayersConnectionHandlerStub.java
@@ -1,6 +1,6 @@
 package com.example.hashcache.stubs;
 
-import com.example.hashcache.database_connections.callbacks.GetPlayerCallback;
+import com.example.hashcache.models.database_connections.callbacks.GetPlayerCallback;
 import com.example.hashcache.models.Player;
 
 import java.util.ArrayList;
